refactor: Move all context related functions to the context file
